feat: finish test and controllers
