+++ app/controllers/consumers/restaurants_controller.rb
@@ -3,11 +3,23 @@ module Consumers
     before_action :authenticate_user!
 
     def index
-      render json: 'ola', status: 200
+      authorize collection
+      render json: RestaurantSerializer.new(collection).serialized_json, status: 200
     end
 
     def show
-      render json: 'ola'
+      authorize restaurant
+      render json: RestaurantSerializer.new(restaurant).serialized_json, status: 200
+    end
+
+    private
+
+    def collection
+      @collection ||= ::Restaurant.all.order(name: :desc)
+    end
+
+    def restaurant
+      @restaurant ||= ::Restaurant.find(params[:id])
     end
   end
 end
